1.修改表结构,default group id  改为1,重新生成代码 2.修改几个项目的pom.xml,将slf4j依赖的范围限制test删除 3.facelog-main 项目实现独立打包

+++ db/src/general/java/net/gdface/facelog/db/DeviceBean.java
@@ -28,7 +28,7 @@ public final class DeviceBean
     private Integer id;
 
     /** comments:所属设备组id */
-    private Integer groupId = new Integer(0)/* DEFAULT:'0'*/;
+    private Integer groupId = new Integer(1)/* DEFAULT:'1'*/;
 
     /** comments:设备名称 */
     private String name;
@@ -202,7 +202,7 @@ public final class DeviceBean
      * <li>full name: fl_device.group_id</li>
      * <li> foreign key: fl_device_group.id</li>
      * <li>comments: 所属设备组id</li>
-     * <li>default value: '0'</li>
+     * <li>default value: '1'</li>
      * <li>column size: 10</li>
      * <li>JDBC type returned by the driver: Types.INTEGER</li>
      * </ul>
