+++ db/src/orm/java/net/gdface/facelog/dborm/device/FlDeviceBean.java
@@ -29,7 +29,7 @@ public  class FlDeviceBean
     private Integer id;
 
     /** comments:所属设备组id */
-    private Integer groupId = new Integer(0)/* DEFAULT:'0'*/;
+    private Integer groupId = new Integer(1)/* DEFAULT:'1'*/;
 
     /** comments:设备名称 */
     private String name;
@@ -195,7 +195,7 @@ public  class FlDeviceBean
      * <li>full name: fl_device.group_id</li>
      * <li> foreign key: fl_device_group.id</li>
      * <li>comments: 所属设备组id</li>
-     * <li>default value: '0'</li>
+     * <li>default value: '1'</li>
      * <li>column size: 10</li>
      * <li>JDBC type returned by the driver: Types.INTEGER</li>
      * </ul>
