+++ facelog-client/src/sql2java/java/net/gdface/facelog/client/DeviceBean.java
@@ -24,7 +24,7 @@ public  class DeviceBean
     private Integer id;
 
     /** comments:所属设备组id */
-    private Integer groupId = new Integer(0)/* DEFAULT:'0'*/;
+    private Integer groupId = new Integer(1)/* DEFAULT:'1'*/;
 
     /** comments:设备名称 */
     private String name;
@@ -190,7 +190,7 @@ public  class DeviceBean
      * <li>full name: fl_device.group_id</li>
      * <li> foreign key: fl_device_group.id</li>
      * <li>comments: 所属设备组id</li>
-     * <li>default value: '0'</li>
+     * <li>default value: '1'</li>
      * <li>column size: 10</li>
      * <li>JDBC type returned by the driver: Types.INTEGER</li>
      * </ul>
